+++ src/components/films-list-extra.js
@@ -1,0 +1,7 @@
+const createFilmsListExtraTemplate = (filmsListTitle) =>
+  `<section class="films-list--extra">
+      <h2 class="films-list__title">${filmsListTitle}</h2>
+      <div class="films-list__container"></div>
+  </section>`;
+
+export {createFilmsListExtraTemplate};
