+++ src/components/films-list.js
@@ -1,0 +1,7 @@
+const createFilmsListTemplate = () =>
+  `<section class="films-list">
+    <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+    <div class="films-list__container"></div>
+  </section>`;
+
+export {createFilmsListTemplate};
