+++ src/components/films.js
@@ -1,0 +1,3 @@
+const createFilmsTemplate = () => `<section class="films"></section>`;
+
+export {createFilmsTemplate};
