+++ src/components/load-more-button.js
@@ -1,0 +1,3 @@
+const createLoadMoreButtonTemplate = () => `<button class="films-list__show-more">Show more</button>`;
+
+export {createLoadMoreButtonTemplate};
